Spliting stamping and creating, adding new method to create
